Change type of UserLocation.accuracy to float (#83)

+++ backend/v1/db/models/user.py
@@ -8,7 +8,7 @@ class UserLocation(BaseModel):
     latitude: float
     longitude: float
     timestamp: Optional[datetime]
-    accuracy: int  # Accuracy in meters
+    accuracy: float  # Accuracy in meters
 
 
 class ShowLocation(str, Enum):
